avance usuarios, productos, colecciones

+++ backend-springboot/src/main/java/com/xoxo/backend/backendspringboot/models/dao/ColeccionDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.xoxo.backend.backendspringboot.models.entities.Coleccion;
 
-public interface ColeccionDao extends CrudRepository<Coleccion, Integer>{
+public interface ColeccionDao extends CrudRepository<Coleccion, Long>{
 
 }
