+++ backend-springboot/src/main/java/com/xoxo/backend/backendspringboot/models/dao/PedidoDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.xoxo.backend.backendspringboot.models.entities.Pedido;
 
-public interface PedidoDao extends CrudRepository<Pedido, Integer>{
+public interface PedidoDao extends CrudRepository<Pedido, Long>{
 
 }
