+++ backend-springboot/src/main/java/com/xoxo/backend/backendspringboot/models/dao/ProductoDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.xoxo.backend.backendspringboot.models.entities.Producto;
 
-public interface ProductoDao extends CrudRepository<Producto, Integer>{
+public interface ProductoDao extends CrudRepository<Producto, Long>{
 
 }
