+++ backend-springboot/src/main/java/com/xoxo/backend/backendspringboot/models/dao/UsuarioDao.java
@@ -3,5 +3,5 @@ package com.xoxo.backend.backendspringboot.models.dao;
 import com.xoxo.backend.backendspringboot.models.entities.Usuario;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsuarioDao extends CrudRepository<Usuario, Integer> {
+public interface UsuarioDao extends CrudRepository<Usuario, Long> {
 }
